chore: self-reference type defs
